backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
asgiref==3.7.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
distlib==0.3.8
dj-database-url==2.1.0
Django==3.2
django-cors-headers==4.3.1
djangorestframework==3.15.1
filelock==3.13.1
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
multidict==6.0.4
mysqlclient==2.2.4
openai==0.27.2
platformdirs==4.2.0
PyMySQL==1.1.0
python-telegram-bot==20.2
pytz==2024.1
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.4
tele==0.0.0
tqdm==4.65.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.15
virtualenv==20.25.0
yarl==1.8.2

